+++ product_portfolio/tests/test_views.py
@@ -34,7 +34,7 @@ from dje.models import Dataspace
 from dje.models import History
 from dje.tasks import logger as tasks_logger
 from dje.tasks import pull_project_data_from_scancodeio
-from dje.tasks import scancodeio_submit_load_sbom
+from dje.tasks import scancodeio_submit_project
 from dje.tests import add_perms
 from dje.tests import create_superuser
 from dje.tests import create_user
@@ -2693,8 +2693,8 @@ class ProductPortfolioViewsTestCase(TestCase):
 
         self.assertEqual(expected, content)
 
-    @mock.patch("dejacode_toolkit.scancodeio.ScanCodeIO.submit_load_sbom")
-    def test_scancodeio_submit_load_sbom_task(self, mock_submit_sbom):
+    @mock.patch("dejacode_toolkit.scancodeio.ScanCodeIO.submit_project")
+    def test_scancodeio_submit_project_task(self, mock_submit_project):
         scancodeproject = ScanCodeProject.objects.create(
             product=self.product1,
             dataspace=self.product1.dataspace,
@@ -2702,15 +2702,16 @@ class ProductPortfolioViewsTestCase(TestCase):
             input_file=ContentFile("Data", name="data.json"),
         )
 
-        mock_submit_sbom.return_value = None
-        scancodeio_submit_load_sbom(
+        mock_submit_project.return_value = None
+        scancodeio_submit_project(
             scancodeproject_uuid=scancodeproject.uuid,
             user_uuid=self.super_user.uuid,
+            pipeline_name="load_sboms",
         )
         scancodeproject.refresh_from_db()
         self.assertEqual("failure", scancodeproject.status)
         self.assertIsNone(scancodeproject.project_uuid)
-        expected = ["- Error: SBOM could not be submitted to ScanCode.io"]
+        expected = ["- Error: File could not be submitted to ScanCode.io"]
         self.assertEqual(expected, scancodeproject.import_log)
 
         # Reset the instance values
@@ -2719,19 +2720,20 @@ class ProductPortfolioViewsTestCase(TestCase):
         scancodeproject.save()
 
         project_uuid = uuid.uuid4()
-        mock_submit_sbom.return_value = {"uuid": project_uuid}
-        scancodeio_submit_load_sbom(
+        mock_submit_project.return_value = {"uuid": project_uuid}
+        scancodeio_submit_project(
             scancodeproject_uuid=scancodeproject.uuid,
             user_uuid=self.super_user.uuid,
+            pipeline_name="load_sboms",
         )
 
         scancodeproject.refresh_from_db()
         self.assertEqual("submitted", scancodeproject.status)
         self.assertEqual(project_uuid, scancodeproject.project_uuid)
-        expected = ["- SBOM file submitted to ScanCode.io for inspection"]
+        expected = ["- File submitted to ScanCode.io for inspection"]
         self.assertEqual(expected, scancodeproject.import_log)
 
-    @mock.patch("dejacode_toolkit.scancodeio.ScanCodeIO.submit_load_sbom")
+    @mock.patch("dejacode_toolkit.scancodeio.ScanCodeIO.submit_project")
     def test_product_portfolio_load_sbom_view(self, mock_submit):
         mock_submit.return_value = None
         self.client.login(username=self.super_user.username, password="secret")
@@ -2743,6 +2745,27 @@ class ProductPortfolioViewsTestCase(TestCase):
         data = {"input_file": ContentFile("Data")}
         response = self.client.post(url, data=data, follow=True)
         expected = "SBOM file submitted to ScanCode.io for inspection."
+        self.assertContains(response, expected)
+        self.assertEqual(1, ScanCodeProject.objects.count())
+
+        with override_settings(CLAMD_ENABLED=True):
+            with mock.patch("dje.fields.SmartFileField.scan_file_for_virus") as scan:
+                data = {"input_file": ContentFile("Data")}
+                self.client.post(url, data=data, follow=True)
+                scan.assert_called_once()
+
+    @mock.patch("dejacode_toolkit.scancodeio.ScanCodeIO.submit_project")
+    def test_product_portfolio_mport_manifest_view(self, mock_submit):
+        mock_submit.return_value = None
+        self.client.login(username=self.super_user.username, password="secret")
+        url = self.product1.get_import_manifests_url()
+        response = self.client.get(url)
+        expected = "Import Packages from manifests"
+        self.assertContains(response, expected)
+
+        data = {"input_file": ContentFile("Data")}
+        response = self.client.post(url, data=data, follow=True)
+        expected = "Manifest file submitted to ScanCode.io for inspection."
         self.assertContains(response, expected)
         self.assertEqual(1, ScanCodeProject.objects.count())